// libs/hdf-converters/types/ionchannel.d.ts
export type ContextualizedDependency = Dependency & {
  parentDependencies: string[];
};
// ...
export type ScanSummary = {
  id: string;
  team_id: string;
  project_id: string;
  analysis_id: string;
  summary: string;
  results: Results;
  created_at: Date;
  updated_at: Date;
  duration: number;
  name: string;
  description: string;
};

export type Results = {
  type: string;
  data: Data;
};

export type Data = {
  vulnerabilities?: DataVulnerability[];
  meta?: Meta;
  dependencies?: Dependency[];
  CSS?: number;
  HTML?: number;
  JavaScript?: number;
  Vue?: number;
  committers?: number;
  name?: string;
  url?: string;
  committed_at?: Date;
  old_names?: string[];
  stars?: number;
  name_changed?: boolean;
  compilers?: null;
  docker_file?: DockerFile;
  known_viruses?: number;
  engine_version?: string;
  scanned_directories?: number;
  scanned_files?: number;
  infected_files?: number;
  data_scanned?: string;
  data_read?: string;
  time?: string;
  file_notes?: Record<string, unknown>;
  clam_av_details?: ClamAVDetails;
  license?: License;
  checksum?: string;
  difference?: boolean;
  message?: string;
  valid?: boolean;
  content?: string;
};

export type ClamAVDetails = {
  clamav_version: string;
  clamav_db_version: string;
};

export type Dependency = {
  latest_version: string;
  org: string;
  name: string;
  type: TypeEnum;
  package: Package;
  version: string;
  scope: Scope;
  requirement: string;
  file: File;
  outdated_version: OutdatedVersion;
  dependencies: Dependency[];
};

export enum File {
  Empty = '',
  PackageJSON = 'package.json'
}

export type OutdatedVersion = {
  major_behind: number;
  minor_behind: number;
  patch_behind: number;
};

export enum Package {
  Package = 'package'
}

export enum Scope {
  Runtime = 'runtime'
}

export enum TypeEnum {
  Npm = 'npm',
  Npmjs = 'npmjs'
}